// backend/packages/server/types/animations.ts
export type AnimationType =
  | "Laughing"
  | "Greet"
  | "Thank"
  | "Sad"
  | "Angry"
  | "Disappointed"
  | "Happy"
  | "Blush"
  | "Dismissing"
  | "Surprised"
  | "Yawn"
  | "Whatever";

export const VALID_ANIMATIONS: AnimationType[] = [
  "Laughing",
  "Greet",
  "Thank",
  "Sad",
  "Angry",
  "Disappointed",
  "Happy",
  "Blush",
  "Dismissing",
  "Surprised",
  "Yawn",
  "Whatever",
];